finish news-page
完成文章顯示, 以及底下的分頁, 但還沒寫Watch(監聽) ，預計抓取資料後才會補上
